Move floor item classes
