+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/contain/Containers.kt
@@ -6,8 +6,8 @@ import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.client.sendContainerItems
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.definition.ContainerDefinitions
-import world.gregs.voidps.engine.entity.item.FloorItems
 import world.gregs.voidps.engine.entity.item.Item
+import world.gregs.voidps.engine.entity.item.floor.FloorItems
 import world.gregs.voidps.engine.utility.get
 
 fun Player.sendContainer(id: String, secondary: Boolean = false) {
