+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/item/floor/FloorItemState.kt
@@ -1,4 +1,4 @@
-package world.gregs.voidps.engine.entity.item
+package world.gregs.voidps.engine.entity.item.floor
 
 sealed class FloorItemState {
     object Private : FloorItemState()
