+++ game/src/main/kotlin/world/gregs/voidps/bot/item/PickupBot.kts
@@ -1,7 +1,7 @@
 import kotlinx.coroutines.CancellableContinuation
 import world.gregs.voidps.engine.entity.Unregistered
 import world.gregs.voidps.engine.entity.contains
-import world.gregs.voidps.engine.entity.item.FloorItem
+import world.gregs.voidps.engine.entity.item.floor.FloorItem
 import world.gregs.voidps.engine.entity.remove
 import world.gregs.voidps.engine.event.on
 import kotlin.coroutines.resume
