+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/item/ItemPlace.kts
@@ -6,7 +6,7 @@ import world.gregs.voidps.engine.entity.World
 import world.gregs.voidps.engine.entity.character.contain.inventory
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.setAnimation
-import world.gregs.voidps.engine.entity.item.FloorItems
+import world.gregs.voidps.engine.entity.item.floor.FloorItems
 import world.gregs.voidps.engine.entity.members
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.engine.map.Distance.getNearest
