+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/effect/AvasDevices.kts
@@ -2,8 +2,8 @@ import world.gregs.voidps.engine.entity.*
 import world.gregs.voidps.engine.entity.character.contain.ItemChanged
 import world.gregs.voidps.engine.entity.character.contain.inventory
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.entity.item.FloorItems
 import world.gregs.voidps.engine.entity.item.equipped
+import world.gregs.voidps.engine.entity.item.floor.FloorItems
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.engine.tick.Job
 import world.gregs.voidps.engine.tick.delay
